# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * The Hive Kenya — offer aggregator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type OfferTier = (typeof OfferTier)[keyof typeof OfferTier];

export const OfferTier = {
  A: "A",
  C: "C",
  D: "D",
} as const;

export type OfferMpesaCompatible =
  (typeof OfferMpesaCompatible)[keyof typeof OfferMpesaCompatible];

export const OfferMpesaCompatible = {
  direct: "direct",
  via_crypto: "via_crypto",
  no: "no",
} as const;

export type OfferSafetyRating =
  (typeof OfferSafetyRating)[keyof typeof OfferSafetyRating];

export const OfferSafetyRating = {
  established: "established",
  new: "new",
  risky: "risky",
} as const;

export type OfferStatus = (typeof OfferStatus)[keyof typeof OfferStatus];

export const OfferStatus = {
  active: "active",
  unconfirmed: "unconfirmed",
  expired: "expired",
} as const;

export interface Offer {
  id: number;
  platformName: string;
  tier: OfferTier;
  depositUsd?: number | null;
  bonusUsd?: number | null;
  lockDays?: number | null;
  withdrawalFeeUsd?: number | null;
  netProfitKes: number;
  hoursRequired?: number | null;
  hourlyRateKes?: number | null;
  activityRequirement?: string | null;
  geographicKenya: boolean;
  mpesaCompatible: OfferMpesaCompatible;
  kycRequired: boolean;
  kycTimeDays?: number | null;
  affiliateLink?: string | null;
  safetyRating: OfferSafetyRating;
  lastVerified?: string | null;
  personalTested: boolean;
  personalNotes?: string | null;
  status: OfferStatus;
  /** Annualised return % for Tier A deposit offers */
  annualisedReturnPct?: number | null;
  /** KES/USD exchange rate used at time of last calculation */
  rateUsedKes?: number | null;
  rateCheckedAt?: string | null;
}

export type SummaryStatsTotalByTier = {
  A: number;
  C: number;
  D: number;
};

export interface SummaryStats {
  totalActiveOffers: number;
  totalByTier: SummaryStatsTotalByTier;
  personalTestedCount: number;
  unconfirmedCount: number;
  highestNetProfitKes: number;
  bestHourlyRateKes: number;
}

export interface TopOffers {
  byNetProfit: Offer[];
  byHourlyRate: Offer[];
}

export interface ErrorResponse {
  error: string;
}

export type ListOffersParams = {
  /**
   * Filter by tier (A=deposit bonus, C=gaming, D=tasks)
   */
  tier?: ListOffersTier;
  /**
   * Filter by M-Pesa compatibility
   */
  mpesa?: ListOffersMpesa;
  /**
   * Sort field
   */
  sort?: ListOffersSort;
};

export type ListOffersTier =
  (typeof ListOffersTier)[keyof typeof ListOffersTier];

export const ListOffersTier = {
  A: "A",
  C: "C",
  D: "D",
} as const;

export type ListOffersMpesa =
  (typeof ListOffersMpesa)[keyof typeof ListOffersMpesa];

export const ListOffersMpesa = {
  direct: "direct",
  via_crypto: "via_crypto",
  no: "no",
} as const;

export type ListOffersSort =
  (typeof ListOffersSort)[keyof typeof ListOffersSort];

export const ListOffersSort = {
  net_profit_kes: "net_profit_kes",
  hourly_rate_kes: "hourly_rate_kes",
  last_verified: "last_verified",
} as const;