# lib/api-zod/src/generated/types/offer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * The Hive Kenya — offer aggregator API
 * OpenAPI spec version: 0.1.0
 */
import type { OfferMpesaCompatible } from "./offerMpesaCompatible";
import type { OfferSafetyRating } from "./offerSafetyRating";
import type { OfferStatus } from "./offerStatus";
import type { OfferTier } from "./offerTier";

export interface Offer {
  id: number;
  platformName: string;
  tier: OfferTier;
  depositUsd?: number | null;
  bonusUsd?: number | null;
  lockDays?: number | null;
  withdrawalFeeUsd?: number | null;
  netProfitKes: number;
  hoursRequired?: number | null;
  hourlyRateKes?: number | null;
  activityRequirement?: string | null;
  geographicKenya: boolean;
  mpesaCompatible: OfferMpesaCompatible;
  kycRequired: boolean;
  kycTimeDays?: number | null;
  affiliateLink?: string | null;
  safetyRating: OfferSafetyRating;
  lastVerified?: Date | null;
  personalTested: boolean;
  personalNotes?: string | null;
  status: OfferStatus;
  /** Annualised return % for Tier A deposit offers */
  annualisedReturnPct?: number | null;
  /** KES/USD exchange rate used at time of last calculation */
  rateUsedKes?: number | null;
  rateCheckedAt?: Date | null;
}