# lib/api-zod/src/generated/types/summaryStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * The Hive Kenya — offer aggregator API
 * OpenAPI spec version: 0.1.0
 */
import type { SummaryStatsTotalByTier } from "./summaryStatsTotalByTier";

export interface SummaryStats {
  totalActiveOffers: number;
  totalByTier: SummaryStatsTotalByTier;
  personalTestedCount: number;
  unconfirmedCount: number;
  highestNetProfitKes: number;
  bestHourlyRateKes: number;
}