# lib/api-zod/src/generated/types/topOffers.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * The Hive Kenya — offer aggregator API
 * OpenAPI spec version: 0.1.0
 */
import type { Offer } from "./offer";

export interface TopOffers {
  byNetProfit: Offer[];
  byHourlyRate: Offer[];
}